[IMPROVEMENT] simplified stakeholder default owner
