+++ components/ILIAS/Exercise/InstructionFile/class.ilExcInstructionFilesStakeholder.php
@@ -27,21 +27,8 @@ use ILIAS\ResourceStorage\Stakeholder\ResourceStakeholder;
  */
 class ilExcInstructionFilesStakeholder extends AbstractResourceStakeholder
 {
-    protected int $owner = 6;
-    private int $current_user;
     protected ?ilDBInterface $database = null;
 
-    /**
-     * ilObjFileStakeholder constructor.
-     */
-    public function __construct(int $owner = 6)
-    {
-        global $DIC;
-        $this->current_user = (int) ($DIC->isDependencyAvailable('user')
-            ? $DIC->user()->getId()
-            : (defined('ANONYMOUS_USER_ID') ? ANONYMOUS_USER_ID : 6));
-        $this->owner = $owner;
-    }
 
     public function getId(): string
     {
@@ -50,7 +37,7 @@ class ilExcInstructionFilesStakeholder extends AbstractResourceStakeholder
 
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner;
+        return $this->default_owner;
     }
 
     public function canBeAccessedByCurrentUser(ResourceIdentification $identification): bool
