+++ components/ILIAS/File/classes/Icons/class.ilObjFileIconStakeholder.php
@@ -27,16 +27,6 @@ use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
  */
 class ilObjFileIconStakeholder extends AbstractResourceStakeholder
 {
-    /**
-     * ilObjFileIconStakeholder constructor.
-     */
-    public function __construct(protected int $owner = 6)
-    {
-    }
-
-    /**
-     * @inheritDoc
-     */
     public function getId(): string
     {
         return 'file_icon';
@@ -47,6 +37,6 @@ class ilObjFileIconStakeholder extends AbstractResourceStakeholder
      */
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner;
+        return $this->default_owner;
     }
 }
