+++ components/ILIAS/File/classes/class.ilObjFileStakeholder.php
@@ -24,17 +24,7 @@ use ILIAS\ResourceStorage\Identification\ResourceIdentification;
  */
 class ilObjFileStakeholder extends AbstractResourceStakeholder
 {
-    private int $current_user;
     protected ?ilDBInterface $database = null;
-
-    /**
-     * ilObjFileStakeholder constructor.
-     */
-    public function __construct(protected int $owner = 6)
-    {
-        global $DIC;
-        $this->current_user = (int) ($DIC->isDependencyAvailable('user') ? $DIC->user()->getId() : ANONYMOUS_USER_ID);
-    }
 
     /**
      * @inheritDoc
@@ -46,7 +36,7 @@ class ilObjFileStakeholder extends AbstractResourceStakeholder
 
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner;
+        return $this->default_owner;
     }
 
     public function canBeAccessedByCurrentUser(ResourceIdentification $identification): bool
