+++ components/ILIAS/FileServices/classes/UploadService/class.ilTemporaryStakeholder.php
@@ -1,5 +1,21 @@
 <?php
 
+/**
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
+ *
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
 use ILIAS\ResourceStorage\Policy\FileNamePolicy;
 use ILIAS\ResourceStorage\Policy\FileNamePolicyException;
 use ILIAS\FileUpload\Processor\BlacklistExtensionPreProcessor;
@@ -16,15 +32,6 @@ use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
  */
 class ilTemporaryStakeholder extends AbstractResourceStakeholder
 {
-    protected int $owner_id;
-
-    public function __construct()
-    {
-        global $DIC;
-
-        $this->owner_id = $DIC->user()->getId();
-    }
-
     public function getId(): string
     {
         return 'irss_temp';
@@ -32,6 +39,6 @@ class ilTemporaryStakeholder extends AbstractResourceStakeholder
 
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner_id;
+        return $this->default_owner;
     }
 }
