+++ components/ILIAS/ILIASObject/classes/Properties/CoreProperties/TileImage/class.ilObjectTileImageStakeholder.php
@@ -21,27 +21,13 @@ declare(strict_types=1);
 namespace ILIAS\Object\Properties\CoreProperties\TileImage;
 
 use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
+use ILIAS\DI\Container;
 
 /**
  * @author Fabian Schmid <fabian@sr.solutions>
  */
 class ilObjectTileImageStakeholder extends AbstractResourceStakeholder
 {
-    private int $default_owner;
-
-    public function __construct()
-    {
-        global $DIC;
-        $this->default_owner = $DIC->isDependencyAvailable('user')
-            ? $DIC->user()->getId()
-            : (defined('SYSTEM_USER_ID') ? (int) SYSTEM_USER_ID : 6);
-    }
-
-    public function setOwner(int $user_id_of_owner): void
-    {
-        $this->default_owner = $user_id_of_owner;
-    }
-
     public function getId(): string
     {
         return 'object_tile_image';
