+++ components/ILIAS/MainMenu/classes/Administration/class.ilMMStorageStakeholder.php
@@ -26,13 +26,6 @@ use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
  */
 class ilMMStorageStakeholder extends AbstractResourceStakeholder
 {
-    public function __construct()
-    {
-    }
-
-    /**
-     * @inheritDoc
-     */
     public function getId(): string
     {
         return 'mme';
@@ -40,6 +33,6 @@ class ilMMStorageStakeholder extends AbstractResourceStakeholder
 
     public function getOwnerOfNewResources(): int
     {
-        return 6;
+        return $this->default_owner;
     }
 }
