+++ components/ILIAS/MetaData/classes/Settings/Copyright/class.ilMDCopyrightImageStakeholder.php
@@ -22,13 +22,6 @@ use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
 
 class ilMDCopyrightImageStakeholder extends AbstractResourceStakeholder
 {
-    protected int $owner = 6;
-
-    public function __construct(int $owner = 6)
-    {
-        $this->owner = $owner;
-    }
-
     public function getId(): string
     {
         return 'copyright_image';
@@ -36,6 +29,6 @@ class ilMDCopyrightImageStakeholder extends AbstractResourceStakeholder
 
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner;
+        return $this->default_owner;
     }
 }
