+++ components/ILIAS/Skill/Profile/class.ilSkillProfileStorageStakeHolder.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 /**
  * This file is part of ILIAS, a powerful learning management system
  * published by ILIAS open source e-Learning e.V.
@@ -16,8 +14,9 @@ declare(strict_types=1);
  * https://www.ilias.de
  * https://github.com/ILIAS-eLearning
  *
- ********************************************************************
- */
+ *********************************************************************/
+
+declare(strict_types=1);
 
 use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
 
@@ -28,26 +27,13 @@ use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
  */
 class ilSkillProfileStorageStakeHolder extends AbstractResourceStakeholder
 {
-    protected int $owner = 6;
-
-    public function __construct(int $owner = 6)
-    {
-        $this->owner = $owner;
-    }
-
-    /**
-     * @inheritDoc
-     */
     public function getId(): string
     {
         return 'skl_prof';
     }
 
-    /**
-     * @inheritDoc
-     */
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner;
+        return $this->default_owner;
     }
 }
