+++ components/ILIAS/StudyProgramme/classes/types/ilStudyProgrammeTypeStakeholder.php
@@ -25,11 +25,6 @@ class ilStudyProgrammeTypeStakeholder extends AbstractResourceStakeholder
 {
     private const ID = 'PRGType';
 
-    public function __construct(
-        protected int $owner = 6
-    ) {
-    }
-
     public function getId(): string
     {
         return self::ID;
@@ -37,7 +32,7 @@ class ilStudyProgrammeTypeStakeholder extends AbstractResourceStakeholder
 
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner;
+        return $this->default_owner;
     }
 
     public function getPreprocessors(): array
