+++ components/ILIAS/WOPI/classes/Handler/WOPIUnknownStakeholder.php
@@ -24,10 +24,6 @@ use ILIAS\ResourceStorage\Stakeholder\AbstractResourceStakeholder;
 
 class WOPIUnknownStakeholder extends AbstractResourceStakeholder
 {
-    public function __construct(protected int $owner = 6)
-    {
-    }
-
     public function getId(): string
     {
         return 'wopi_unknown';
@@ -35,7 +31,7 @@ class WOPIUnknownStakeholder extends AbstractResourceStakeholder
 
     public function getOwnerOfNewResources(): int
     {
-        return $this->owner;
+        return $this->default_owner;
     }
 
 }
